Minor refactoring, moving classes to correct files
