+++ src/Nether.Analytics.EventProcessor/Nether.Analytics.EventProcessor.csproj
@@ -105,6 +105,9 @@
     <Compile Include="ConfigResolver.cs" />
     <Compile Include="ConsoleEx.cs" />
     <Compile Include="GameEventData.cs" />
+    <Compile Include="GameEvents\IncommingLocationEvent.cs" />
+    <Compile Include="GameEvents\IOutGameEvent.cs" />
+    <Compile Include="GameEvents\OutgoingLocationEvent.cs" />
     <Compile Include="NameResolver.cs" />
     <Compile Include="Output\Blob\BlobOutputManager.cs" />
     <Compile Include="Output\EventHub\EventHubOutputManager.cs" />
@@ -112,7 +115,6 @@
     <Compile Include="EventProcessorExtensions.cs" />
     <Compile Include="GameEventReceiver.cs" />
     <Compile Include="GameEventRouter.cs" />
-    <Compile Include="GameEvents\OutLocationEvent.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringBuilderExtensions.cs" />
